x/tokenmanager/client/cli/query_item.go
package cli

import (
	"context"

	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/spf13/cobra"
	"gitlab.com/rarify-protocol/rarimo-core/x/tokenmanager/types"
)

func CmdListItem() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "list-item",
		Short: "list all item",
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			pageReq, err := client.ReadPageRequest(cmd.Flags())
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryAllItemRequest{
				Pagination: pageReq,
			}

			res, err := queryClient.ItemAll(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}

func CmdShowItem() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-item [token-address] [token-id]",
		Short: "shows a item",
		Args:  cobra.ExactArgs(2),
		RunE: func(cmd *cobra.Command, args []string) (err error) {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			argTokenAddress := args[0]
			argTokenId := args[1]

			params := &types.QueryGetItemRequest{
				TokenAddress: argTokenAddress,
				TokenId:      argTokenId,
			}

			res, err := queryClient.Item(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
